Restructure database schema and tag api added

+++ Server/src/main/java/com/unishare/backend/DTO/Request/RegisterRequest.java
@@ -39,7 +39,7 @@ public class RegisterRequest {
     private double lng;
     @NotNull
     @NotEmpty
-    private Integer university;
+    private Long university;
     @NotNull
     @NotEmpty
     private Role role;
